Stocker/requirements.txt
adhoc-python==1.0.2
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
args==0.1.0
asgiref==3.7.2
attrs==23.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
better-profanity==0.7.0
billiard==4.2.1
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clint==0.5.1
colorama==0.4.6
comtypes==1.4.2
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.1
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==4.0.0
django-jazzmin==2.6.0
django-location-field==2.7.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
dnspython==2.7.0
ecapture==2.0.2
email_validator==2.2.0
falcon==3.1.3
fastapi==0.115.3
fastapi-cli==0.0.5
feedparser==6.0.11
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
jaraco.context==5.3.0
jasmin==0.11.1
Jinja2==3.1.4
kombu==5.4.2
lockfile==0.12.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
opencv-python==4.9.0.80
outcome==1.3.0.post0
Pillow==9.5.0
prometheus-client==0.18.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyjokes==0.6.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.4
pypiwin32==223
pypng==0.20220715.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
python-messaging==0.5.13
python-mimeparse==1.6.0
python-multipart==0.0.12
python3-openid==3.2.0
pyttsx3==2.90
pytz==2024.2
pywin32==306
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
selenium==4.24.0
service-identity==23.1.0
setuptools==70.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smpp.pdu3==0.6
smpp.twisted3==0.8
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
sqlparse==0.4.4
starlette==0.41.0
tabulate==0.8.10
treq==23.11.0
trio==0.26.2
trio-websocket==0.11.1
twilio==9.1.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txAMQP3==0.9.4
txredisapi==1.4.10
typer==0.12.5
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.32.0
vine==5.1.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.4
zope.interface==7.1.1
